Add support for using OAuth2 refresh tokens.
The Tastytrade.from_env convenience method has been added to load OAuth2 parameters from the environment, but these can also be passed directly to the constructor.

This additionally deprecates support for login via session.

+++ src/tastytrade_sdk/tastytrade.py
@@ -1,3 +1,6 @@
+from os import environ
+from typing import Optional
+import warnings
 from injector import Injector
 
 from tastytrade_sdk.config import Config
@@ -10,20 +13,30 @@ class Tastytrade:
     The SDK's top-level class
     """
 
-    def __init__(self, api_base_url: str = 'api.tastytrade.com'):
+    def __init__(self,
+                 api_base_url: str = 'api.tastytrade.com',
+                 client_id: Optional[str] = None,
+                 client_secret: Optional[str] = None,
+                 refresh_token: Optional[str] = None):
         """
         :param api_base_url: Optionally override the base URL used by the API
         (when using the sandbox environment, for e.g.)
         """
 
         def __configure(binder):
-            binder.bind(Config, to=Config(api_base_url=api_base_url))
+            binder.bind(Config, to=Config(api_base_url=api_base_url,
+                                          client_id=client_id,
+                                          client_secret=client_secret,
+                                          refresh_token=refresh_token))
 
         self.__container = Injector(__configure)
 
     def login(self, login: str, password: str) -> 'Tastytrade':
         """
         Initialize a logged-in session
+
+        .. deprecated:: 1.3.0
+            Use OAuth2 credentials instead.
         """
         self.__container.get(RequestsSession).login(login, password)
         return self
@@ -31,8 +44,30 @@ class Tastytrade:
     def logout(self) -> None:
         """
         End the session
+
+        .. deprecated:: 1.3.0
+            Use OAuth2 credentials instead.
         """
-        self.api.delete('/sessions')
+        if self.__container.get(RequestsSession).is_oauth_session:
+            warnings.warn('Logout does not need to be called with an OAuth2 session.', DeprecationWarning)
+        else:
+            self.api.delete('/sessions')
+
+    @classmethod
+    def from_env(cls) -> 'Tastytrade':
+        """Creates a Tastytrade OAuth2 session using the environment variables:
+    
+        - `API_BASE_URL` (uses `api.tastyworks.com` unless specified)
+        - `TT_CLIENT_ID`
+        - `TT_CLIENT_SECRET`
+        - `TT_REFRESH_TOKEN`
+        """
+        return Tastytrade(api_base_url=environ.get('API_BASE_URL', 'api.tastyworks.com'),
+                          client_id=environ.get('TT_CLIENT_ID'),
+                          client_secret=environ.get('TT_CLIENT_SECRET'),
+                          refresh_token=environ.get('TT_REFRESH_TOKEN')
+        )
+
 
     @property
     def market_data(self) -> MarketData:
